requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.8.4
torchmetrics==0.11.0

# --------- hydra --------- #
hydra-core==1.3.1
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils
pre-commit
rich
pytest
sh; platform_system != 'Windows'
omegaconf
tqdm
tensorflow
pynvml

# --------- data --------- #
numpy
Pillow
opencv-python-headless
albumentations
h5py

# --------- models --------- #
timm
segmentation-models-pytorch
monai
segment_anything

# --------- visualization --------- #
seaborn
scikit_learn
matplotlib
ipython
grad-cam==1.3.1
ttach==0.0.3
kaggle
